+++ src/crc32.h
@@ -7,6 +7,6 @@
 #define CRC32_POLYNOMIAL	0xEDB88320L
 
 int build_crc_table(void);
-unsigned long compute_crc32(unsigned char *, int);
+unsigned /*long*/ int compute_crc32(unsigned char *, int);
 void free_crc_table(void);
