+++ src/loader_id.c
@@ -40,11 +40,11 @@
  * The crc passed should obviously be that of the first CBM program in the TAP.
  */
 
-int idloader(unsigned long crc)
+int idloader(unsigned /*long*/ int crc)
 {
 	int i, id, len;
 
-	unsigned long kcrc[NUM_CRC][2] = {
+	unsigned /*long*/ int kcrc[NUM_CRC][2] = {
 		{0x261DBA0E, LID_FREE},
 		{0x8AAE883E, LID_FREE},
 		{0x2EF78649, LID_FREE},
